requirements.txt
Flask==3.1.2
gunicorn
python-dotenv
replicate

